+++ tests/test_cli.py
@@ -1,5 +1,7 @@
-from click.testing import CliRunner
+import os
 import sys
+import pytest
+from click.testing import CliRunner
 
 PY2 = sys.version_info[0] == 2
 
@@ -54,26 +56,23 @@ def test_show__format_html():
 def test_show__output_to_file():
     from pycobertura.cli import show
 
-    with open('tests/dummy.original.xml') as f:
-        cobertura_content = f.read()
-
     runner = CliRunner()
     for opt in ('-o', '--output'):
-        with runner.isolated_filesystem():
-            with open('cobertura.xml', 'w') as f:
-                f.write(cobertura_content)
-            result = runner.invoke(show, [
-                'cobertura.xml', opt, 'report.out'
-            ], catch_exceptions=False)
-            with open('report.out') as f:
-                report = f.read()
-            assert result.output == ""
-            assert report == """\
-Name              Stmts    Miss  Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__        0       0  0.00%
-dummy/dummy           4       2  50.00%   2, 5
-TOTAL                 4       2  50.00%"""
+        result = runner.invoke(show, [
+            'tests/cobertura.xml', opt, 'report.out'
+        ], catch_exceptions=False)
+        with open('report.out') as f:
+            report = f.read()
+        os.remove('report.out')
+        assert result.output == ""
+        assert report == """\
+Name                         Stmts    Miss  Cover    Missing
+-------------------------  -------  ------  -------  ---------
+Main                            11       0  100.00%
+search.BinarySearch             12       1  91.67%   24
+search.ISortedArraySearch        0       0  100.00%
+search.LinearSearch              7       2  71.43%   19-24
+TOTAL                           30       3  90.00%"""
 
 
 def test_diff__format_default():
@@ -81,27 +80,29 @@ def test_diff__format_default():
 
     runner = CliRunner()
     result = runner.invoke(diff, [
-        'tests/dummy.with-dummy2-better-cov.xml',
-        'tests/dummy.with-dummy2-better-and-worse.xml',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source2/coverage.xml',
     ], catch_exceptions=False)
     # FIXME: Fails in PY2, requires click>=4.x to pass
     if not PY2:
         assert result.output == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  \x1b[31m+5\x1b[39m
-dummy/dummy2    -        -1      +50.00%  \x1b[32m-2\x1b[39m
-TOTAL           -        -       -
+dummy/dummy     -        -2      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
+dummy/dummy2    +2       +1      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
+dummy/dummy3    +2       +2      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
+TOTAL           -1       -4      +31.06%
 """
     else:
         assert result.output == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +5
-dummy/dummy2    -        -1      +50.00%  -2
-TOTAL           -        -       -
+dummy/dummy     -        -2      +40.00%  -5, -6
+dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
+dummy/dummy3    +2       +2      -        +1, +2
+TOTAL           -1       -4      +31.06%
 """
 
 
@@ -112,65 +113,55 @@ def test_diff__format_text():
     for opt in ('-f', '--format'):
         result = runner.invoke(diff, [
             opt, 'text',
-            'tests/dummy.with-dummy2-better-cov.xml',
-            'tests/dummy.with-dummy2-better-and-worse.xml',
+            'tests/dummy.source1/coverage.xml',
+            'tests/dummy.source2/coverage.xml',
         ], catch_exceptions=False)
         # FIXME: Fails in PY2, requires click>=4.x to pass
         if not PY2:
             assert result.output == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  \x1b[31m+5\x1b[39m
-dummy/dummy2    -        -1      +50.00%  \x1b[32m-2\x1b[39m
-TOTAL           -        -       -
+dummy/dummy     -        -2      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
+dummy/dummy2    +2       +1      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
+dummy/dummy3    +2       +2      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
+TOTAL           -1       -4      +31.06%
 """
         else:
             assert result.output == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +5
-dummy/dummy2    -        -1      +50.00%  -2
-TOTAL           -        -       -
+dummy/dummy     -        -2      +40.00%  -5, -6
+dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
+dummy/dummy3    +2       +2      -        +1, +2
+TOTAL           -1       -4      +31.06%
 """
 
 
 def test_diff__output_to_file():
     from pycobertura.cli import diff
 
-    with open('tests/dummy.with-dummy2-better-cov.xml') as f:
-        cobertura1_content = f.read()
-
-    with open('tests/dummy.with-dummy2-better-and-worse.xml') as f:
-        cobertura2_content = f.read()
-
     runner = CliRunner()
 
     for opt in ('-o', '--output'):
-        with runner.isolated_filesystem():
-
-            with open('cobertura1.xml', 'w') as f:
-                f.write(cobertura1_content)
-
-            with open('cobertura2.xml', 'w') as f:
-                f.write(cobertura2_content)
-
-            result = runner.invoke(diff, [
-                'cobertura1.xml',
-                'cobertura2.xml',
-                opt, 'report.out'
-            ], catch_exceptions=False)
-            with open('report.out') as f:
-                report = f.read()
-            assert result.output == ""
-            assert report == """\
+        result = runner.invoke(diff, [
+            'tests/dummy.source1/coverage.xml',
+            'tests/dummy.source2/coverage.xml',
+            opt, 'report.out'
+        ], catch_exceptions=False)
+        with open('report.out') as f:
+            report = f.read()
+        os.remove('report.out')
+        assert result.output == ""
+        assert report == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +5
-dummy/dummy2    -        -1      +50.00%  -2
-TOTAL           -        -       -"""
+dummy/dummy     -        -2      +40.00%  -5, -6
+dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
+dummy/dummy3    +2       +2      -        +1, +2
+TOTAL           -1       -4      +31.06%"""
 
 
 # FIXME: when Click 4 is available, uncomment this.
@@ -192,7 +183,18 @@ TOTAL           -        -       -"""
 #"""
 
 
-def test_diff__format_html():
+def test_diff__format_html__no_source():
+    from pycobertura.cli import diff
+
+    runner = CliRunner()
+    pytest.raises(IOError, runner.invoke, diff, [
+        '--format', 'html',
+        'tests/dummy.with-dummy2-better-cov.xml',
+        'tests/dummy.with-dummy2-better-and-worse.xml',
+    ], catch_exceptions=False)
+
+
+def test_diff__format_html__with_source():
     from pycobertura.cli import diff
 
     runner = CliRunner()
@@ -207,13 +209,57 @@ def test_diff__format_html():
     assert result.output.endswith('</html>\n')
 
 
+def test_diff__format_html__missed():
+    from pycobertura.cli import diff
+
+    runner = CliRunner()
+    result = runner.invoke(diff, [
+        '--format', 'html',
+        '--missed',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source2/coverage.xml',
+    ], catch_exceptions=False)
+    assert 'Missing' in result.output
+    assert result.output.startswith('<html>')
+    assert result.output.endswith('</html>\n')
+
+
+def test_diff__format_html__missed_is_default():
+    from pycobertura.cli import diff
+
+    runner = CliRunner()
+    result = runner.invoke(diff, [
+        '--format', 'html',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source2/coverage.xml',
+    ], catch_exceptions=False)
+    assert 'Missing' in result.output
+    assert result.output.startswith('<html>')
+    assert result.output.endswith('</html>\n')
+
+
+def test_diff__format_html__no_missed():
+    from pycobertura.cli import diff
+
+    runner = CliRunner()
+    result = runner.invoke(diff, [
+        '--format', 'html',
+        '--no-missed',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source2/coverage.xml',
+    ], catch_exceptions=False)
+    assert 'Missing' not in result.output
+    assert result.output.startswith('<html>')
+    assert result.output.endswith('</html>\n')
+
+
 def test_diff__same_coverage_has_exit_status_of_zero():
     from pycobertura.cli import diff
 
     runner = CliRunner()
     result = runner.invoke(diff, [
-        'tests/dummy.original.xml',
-        'tests/dummy.original.xml',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source1/coverage.xml',
     ], catch_exceptions=False)
     assert result.exit_code == 0
 
@@ -223,8 +269,8 @@ def test_diff__better_coverage_has_exit_status_of_zero():
 
     runner = CliRunner()
     result = runner.invoke(diff, [
-        'tests/dummy.original.xml',
-        'tests/dummy.original-full-cov.xml',
+        'tests/dummy.source1/coverage.xml',
+        'tests/dummy.source2/coverage.xml',
     ], catch_exceptions=False)
     assert result.exit_code == 0
 
@@ -234,7 +280,7 @@ def test_diff__worse_coverage_has_exit_status_of_one():
 
     runner = CliRunner()
     result = runner.invoke(diff, [
-        'tests/dummy.original-full-cov.xml',
-        'tests/dummy.original.xml',
+        'tests/dummy.source2/coverage.xml',
+        'tests/dummy.source1/coverage.xml',
     ], catch_exceptions=False)
     assert result.exit_code == 1