+++ tests/test_reporters.py
@@ -12,24 +12,6 @@ def remove_style_tag(html):
     return html_nostyle
 
 
-def test_report_row_by_class():
-    from pycobertura.reporters import TextReporter
-
-    cobertura = make_cobertura()
-    report = TextReporter(cobertura)
-
-    expected_rows = {
-        'Main': ['Main', 11, 0, 1, []],
-        'search.BinarySearch': ['search.BinarySearch', 12, 1, 0.9166666666666666, [24]],
-        'search.ISortedArraySearch': ['search.ISortedArraySearch', 0, 0, 1.0, []],
-        'search.LinearSearch': ['search.LinearSearch', 7, 2, 0.7142857142857143, [19, 20, 21, 22, 23, 24]],
-    }
-
-    for class_name in cobertura.classes():
-        row = report.get_report_row_by_class(class_name)
-        assert row == expected_rows[class_name]
-
-
 def test_text_report():
     from pycobertura.reporters import TextReporter
 
@@ -64,8 +46,8 @@ TOTAL                 6       2  66.67%"""
 def test_text_report_delta__no_diff():
     from pycobertura.reporters import TextReporter, TextReporterDelta
 
-    cobertura1 = make_cobertura('tests/dummy.original.xml')
-    cobertura2 = make_cobertura('tests/dummy.original.xml')
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source1/coverage.xml')
 
     report_delta = TextReporterDelta(cobertura1, cobertura2)
 
@@ -74,208 +56,63 @@ Name            Stmts    Miss    Cover    Missing
 --------------  -------  ------  -------  ---------
 dummy/__init__  -        -       -
 dummy/dummy     -        -       -
-TOTAL           -        -       -"""
-
-
-def test_text_report_delta__improve_coverage():
-    from pycobertura.reporters import TextReporter, TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original.xml')
-    cobertura2 = make_cobertura('tests/dummy.original-better-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -1      +25.00%  -2
-TOTAL           -        -1      +25.00%"""
-
-
-def test_text_report_delta__full_coverage():
-    from pycobertura.reporters import TextReporter, TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original.xml')
-    cobertura2 = make_cobertura('tests/dummy.original-full-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -2      +50.00%  -2, -5
-TOTAL           -        -2      +50.00%"""
-
-
-def test_text_report_delta__worsen_coverage():
-    from pycobertura.reporters import TextReporter, TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original-better-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.original.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +2
-TOTAL           -        +1      -25.00%"""
-
-
-def test_text_report_delta__new_dummy2_class_has_no_coverage():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original-full-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-no-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -       -
-dummy/dummy2    +2       +2      -        +1, +2
-TOTAL           +2       +2      -33.33%"""
-
-
-def test_text_report_delta__new_dummy2_class_has_improved_coverage():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original-full-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-better-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -       -
-dummy/dummy2    +2       +1      +50.00%  +2
-TOTAL           +2       +1      -16.67%"""
-
-
-def test_text_report_delta__new_dummy2_class_has_full_coverage():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.original-full-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-full-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover     Missing
---------------  -------  ------  --------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -       -
-dummy/dummy2    +2       -       +100.00%
-TOTAL           +2       -       -"""
-
-
-def test_text_report_delta__removed_dummy2_class():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.with-dummy2-full-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.original-full-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        -       -
-dummy/dummy2    -2       -       -
-TOTAL           -2       -       -"""
-
-
-def test_text_report_delta__removed_dummy2_class_and_worsen_coverage():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.with-dummy2-full-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.original-better-cov.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +5
-dummy/dummy2    -2       -       -
-TOTAL           -2       +1      -25.00%"""
-
-
-def test_text_report_delta__better_and_worse_coverage():
-    from pycobertura.reporters import TextReporterDelta
-
-    cobertura1 = make_cobertura('tests/dummy.with-dummy2-better-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-better-and-worse.xml')
-
-    report_delta = TextReporterDelta(cobertura1, cobertura2)
-
-    assert report_delta.generate() == """\
-Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
-dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  +5
-dummy/dummy2    -        -1      +50.00%  -2
+dummy/dummy2    -        -       -
+dummy/dummy4    -        -       -
 TOTAL           -        -       -"""
 
 
 def test_text_report_delta__colorize_True():
     from pycobertura.reporters import TextReporterDelta
 
-    cobertura1 = make_cobertura('tests/dummy.with-dummy2-better-cov.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-better-and-worse.xml')
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
 
     report_delta = TextReporterDelta(cobertura1, cobertura2, color=True)
 
     assert report_delta.generate() == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        +1      -25.00%  \x1b[31m+5\x1b[39m
-dummy/dummy2    -        -1      +50.00%  \x1b[32m-2\x1b[39m
-TOTAL           -        -       -"""
+dummy/dummy     -        -2      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
+dummy/dummy2    +2       +1      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
+dummy/dummy3    +2       +2      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
+TOTAL           -1       -4      +31.06%"""
 
 
 def test_text_report_delta__colorize_True__with_missing_range():
     from pycobertura.reporters import TextReporterDelta
 
-    cobertura1 = make_cobertura('tests/dummy.original.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-no-cov.xml')
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
 
     report_delta = TextReporterDelta(cobertura1, cobertura2, color=True)
 
     assert report_delta.generate() == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        -2      +50.00%  \x1b[32m-2\x1b[39m, \x1b[32m-5\x1b[39m
-dummy/dummy2    +2       +2      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
-TOTAL           +2       -       +16.67%"""
+dummy/dummy     -        -2      +40.00%  \x1b[32m-5\x1b[39m, \x1b[32m-6\x1b[39m
+dummy/dummy2    +2       +1      -25.00%  \x1b[32m-2\x1b[39m, \x1b[32m-4\x1b[39m, \x1b[31m+5\x1b[39m
+dummy/dummy3    +2       +2      -        \x1b[31m+1\x1b[39m, \x1b[31m+2\x1b[39m
+TOTAL           -1       -4      +31.06%"""
 
 
 def test_text_report_delta__colorize_False():
     from pycobertura.reporters import TextReporterDelta
 
-    cobertura1 = make_cobertura('tests/dummy.original.xml')
-    cobertura2 = make_cobertura('tests/dummy.with-dummy2-no-cov.xml')
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
 
     report_delta = TextReporterDelta(cobertura1, cobertura2, color=False)
 
     assert report_delta.generate() == """\
 Name            Stmts    Miss    Cover    Missing
---------------  -------  ------  -------  ---------
+--------------  -------  ------  -------  ----------
 dummy/__init__  -        -       -
-dummy/dummy     -        -2      +50.00%  -2, -5
-dummy/dummy2    +2       +2      -        +1, +2
-TOTAL           +2       -       +16.67%"""
+dummy/dummy     -        -2      +40.00%  -5, -6
+dummy/dummy2    +2       +1      -25.00%  -2, -4, +5
+dummy/dummy3    +2       +2      -        +1, +2
+TOTAL           -1       -4      +31.06%"""
 
 
 def test_html_report():
@@ -353,7 +190,7 @@ def test_html_report():
 </pre>
   </div>
   <div class="eleven columns code">
-    <pre><span class="noop">Main.java not found</span></pre>
+    <pre><span class="noop">tests/Main.java not found</span></pre>
   </div>
 </div>
 <h4 id="search.BinarySearch">search.BinarySearch</h4>
@@ -363,7 +200,7 @@ def test_html_report():
 </pre>
   </div>
   <div class="eleven columns code">
-    <pre><span class="noop">search/BinarySearch.java not found</span></pre>
+    <pre><span class="noop">tests/search/BinarySearch.java not found</span></pre>
   </div>
 </div>
 <h4 id="search.ISortedArraySearch">search.ISortedArraySearch</h4>
@@ -373,7 +210,7 @@ def test_html_report():
 </pre>
   </div>
   <div class="eleven columns code">
-    <pre><span class="noop">search/ISortedArraySearch.java not found</span></pre>
+    <pre><span class="noop">tests/search/ISortedArraySearch.java not found</span></pre>
   </div>
 </div>
 <h4 id="search.LinearSearch">search.LinearSearch</h4>
@@ -383,9 +220,95 @@ def test_html_report():
 </pre>
   </div>
   <div class="eleven columns code">
-    <pre><span class="noop">search/LinearSearch.java not found</span></pre>
+    <pre><span class="noop">tests/search/LinearSearch.java not found</span></pre>
   </div>
 </div>
+    </div>
+  </body>
+</html>"""
+
+
+def test_text_report_delta__no_missed():
+    from pycobertura.reporters import TextReporterDelta
+
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
+
+    report_delta = TextReporterDelta(cobertura1, cobertura2, show_missed=False)
+    output = report_delta.generate()
+    assert output == """\
+Name            Stmts    Miss    Cover
+--------------  -------  ------  -------
+dummy/__init__  -        -       -
+dummy/dummy     -        -2      +40.00%
+dummy/dummy2    +2       +1      -25.00%
+dummy/dummy3    +2       +2      -
+TOTAL           -1       -4      +31.06%"""
+
+
+def test_html_report_delta__no_missed():
+    from pycobertura.reporters import HtmlReporterDelta
+
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
+
+    report_delta = HtmlReporterDelta(cobertura1, cobertura2, show_missed=False)
+    html_output = report_delta.generate()
+    assert 'Missing' not in html_output
+    assert '<h4 id=' not in html_output
+
+    assert remove_style_tag(html_output) == """\
+<html>
+  <head>
+    <title>pycobertura report</title>
+    <meta charset="UTF-8">
+  </head>
+  <body>
+    <div class="container">
+      <table class="u-full-width">
+        <thead>
+          <tr>
+            <th>Name</th>
+            <th>Stmts</th>
+            <th>Miss</th>
+            <th>Cover</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td><a href="#dummy/__init__">dummy/__init__</a></td>
+            <td>-</td>
+            <td>-</td>
+            <td>-</td>
+          </tr>
+          <tr>
+            <td><a href="#dummy/dummy">dummy/dummy</a></td>
+            <td>-</td>
+            <td>-2</td>
+            <td>+40.00%</td>
+          </tr>
+          <tr>
+            <td><a href="#dummy/dummy2">dummy/dummy2</a></td>
+            <td>+2</td>
+            <td>+1</td>
+            <td>-25.00%</td>
+          </tr>
+          <tr>
+            <td><a href="#dummy/dummy3">dummy/dummy3</a></td>
+            <td>+2</td>
+            <td>+2</td>
+            <td>-</td>
+          </tr>
+        </tbody>
+        <tfoot>
+          <tr>
+            <td>TOTAL</td>
+            <td>-1</td>
+            <td>-4</td>
+            <td>+31.06%</td>
+          </tr>
+        </tfoot>
+      </table>
     </div>
   </body>
 </html>"""
@@ -394,8 +317,8 @@ def test_html_report():
 def test_html_report_delta():
     from pycobertura.reporters import HtmlReporterDelta
 
-    cobertura1 = make_cobertura('tests/dummy.source1.xml', base_path='tests/dummy.source1')
-    cobertura2 = make_cobertura('tests/dummy.source2.xml', base_path='tests/dummy.source2')
+    cobertura1 = make_cobertura('tests/dummy.source1/coverage.xml')
+    cobertura2 = make_cobertura('tests/dummy.source2/coverage.xml')
 
     report_delta = HtmlReporterDelta(cobertura1, cobertura2)
     html_output = report_delta.generate()
@@ -444,7 +367,7 @@ def test_html_report_delta():
             <td>+2</td>
             <td>+1</td>
             <td>-25.00%</td>
-            <td><span class="red">+5</span>
+            <td><span class="green">-2</span>, <span class="green">-4</span>, <span class="red">+5</span>
             </td>
           </tr>
           <tr>
@@ -453,14 +376,6 @@ def test_html_report_delta():
             <td>+2</td>
             <td>-</td>
             <td><span class="red">+1</span>, <span class="red">+2</span>
-            </td>
-          </tr>
-          <tr>
-            <td><a href="#dummy/dummy4">dummy/dummy4</a></td>
-            <td>-5</td>
-            <td>-</td>
-            <td>-</td>
-            <td>
             </td>
           </tr>
         </tbody>